+++ src/commonMain/kotlin/util/Either.kt
@@ -54,7 +54,7 @@ sealed class Either<out L, out R> {
      * Deprecated in 1.3. [Either] is right-biased.
      * All methods from [RightProjection] should be called directly on [Either].
      */
-    @Suppress("DeprecatedCallableAddReplaceWith")
+    @Suppress("DeprecatedCallableAddReplaceWith", "DEPRECATION")
     @Deprecated(
         message = "Either is right-biased. All methods from `RightProjection` should be called directly on `Either`.",
         level = DeprecationLevel.WARNING
@@ -357,7 +357,7 @@ fun <L, R> Either<L, R?>.filterNotNullToOption(): Option<Either<L, R>> = when (t
  *
  * @since 1.3
  */
-inline fun <L, R> Either<L, R>.filterOrElse(predicate: (R) -> Boolean, zero: () -> L): Either<L, R>? =
+inline fun <L, R> Either<L, R>.filterOrElse(predicate: (R) -> Boolean, zero: () -> L): Either<L, R> =
     when (this) {
         is Right -> if (predicate(value)) this else Left(zero())
         is Left -> this
@@ -656,7 +656,7 @@ fun <L, R> LeftProjection<L?, R>.filterNotNullToOption(): Option<Either<L, R>> =
  *
  * @since 1.2
  */
-inline fun <L, R> LeftProjection<L, R>.filterOrElse(predicate: (L) -> Boolean, zero: () -> R): Either<L, R>? =
+inline fun <L, R> LeftProjection<L, R>.filterOrElse(predicate: (L) -> Boolean, zero: () -> R): Either<L, R> =
     when (either) {
         is Left -> if (predicate(either.value)) either else Right(zero())
         is Right -> either
@@ -848,6 +848,7 @@ data class RightProjection<out L, out R>(val either: Either<L, R>) {
  *
  * @return Value of this [Right] or [default].
  */
+@Suppress("DEPRECATION")
 inline fun <L, R> RightProjection<L, R>.getOrElse(default: () -> R): R = when (either) {
     is Left -> default()
     is Right -> either.value
@@ -860,6 +861,7 @@ inline fun <L, R> RightProjection<L, R>.getOrElse(default: () -> R): R = when (e
  *
  * @return [Either] mapped using [transform] or this object if this is a [Left].
  */
+@Suppress("DEPRECATION")
 inline fun <L, R, T> RightProjection<L, R>.flatMap(transform: (R) -> Either<L, T>): Either<L, T> = when (either) {
     is Left -> either
     is Right -> transform(either.value)
@@ -870,6 +872,7 @@ inline fun <L, R, T> RightProjection<L, R>.flatMap(transform: (R) -> Either<L, T
  *
  * @return The same [Right] if its value is not `null`. Otherwise returns `null`.
  */
+@Suppress("DEPRECATION")
 fun <L, R> RightProjection<L, R?>.filterNotNull(): Either<L, R>? = when (either) {
     is Left -> null
     is Right -> either.value?.let { Right(it) }
@@ -880,6 +883,7 @@ fun <L, R> RightProjection<L, R?>.filterNotNull(): Either<L, R>? = when (either)
  *
  * @return [Some] containing the same [Right] if its value is not `null`. Otherwise returns [None].
  */
+@Suppress("DEPRECATION")
 fun <L, R> RightProjection<L, R?>.filterNotNullToOption(): Option<Either<L, R>> = when (either) {
     is Left -> None
     is Right -> either.value?.let { Right(it) }.asOption()
@@ -899,7 +903,8 @@ fun <L, R> RightProjection<L, R?>.filterNotNullToOption(): Option<Either<L, R>> 
  *
  * @since 1.2
  */
-inline fun <L, R> RightProjection<L, R>.filterOrElse(predicate: (R) -> Boolean, zero: () -> L): Either<L, R>? =
+@Suppress("DEPRECATION")
+inline fun <L, R> RightProjection<L, R>.filterOrElse(predicate: (R) -> Boolean, zero: () -> L): Either<L, R> =
     when (either) {
         is Right -> if (predicate(either.value)) either else Left(zero())
         is Left -> either